fix(css): align local ident hash with webpack
