+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/fullhash-local.txt
@@ -1,8 +1,8 @@
 Object {
-  btn--info_is-disabled_1: _1a2aa831eabab351-btn--info_is-disabled_1,
-  btn-info_is-disabled: _150bd62048895dcd-btn-info_is-disabled,
+  btn--info_is-disabled_1: eKO8qT-btn--info_is-disabled_1,
+  btn-info_is-disabled: fq80i9-btn-info_is-disabled,
   foo: bar,
-  foo_bar: _256451fed46ecd88-foo_bar,
+  foo_bar: BgYTlD-foo_bar,
   my-btn-info_is-disabled: value,
-  simple: _1e0587bd670b0b5b-simple,
+  simple: xmLu3J-simple,
 }
