+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/hash-local-custom-node.txt
@@ -1,0 +1,8 @@
+Object {
+  btn--info_is-disabled_1: b24c48185173635cf9a1-btn--info_is-disabled_1,
+  btn-info_is-disabled: b24c48185173635cf9a1-btn-info_is-disabled,
+  foo: bar,
+  foo_bar: b24c48185173635cf9a1-foo_bar,
+  my-btn-info_is-disabled: value,
+  simple: b24c48185173635cf9a1-simple,
+}
