+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/hash-local-custom.txt
@@ -1,8 +1,8 @@
 Object {
-  btn--info_is-disabled_1: deadd45e622d3803-btn--info_is-disabled_1,
-  btn-info_is-disabled: deadd45e622d3803-btn-info_is-disabled,
+  btn--info_is-disabled_1: f7100025dd3660945f22-btn--info_is-disabled_1,
+  btn-info_is-disabled: f7100025dd3660945f22-btn-info_is-disabled,
   foo: bar,
-  foo_bar: deadd45e622d3803-foo_bar,
+  foo_bar: f7100025dd3660945f22-foo_bar,
   my-btn-info_is-disabled: value,
-  simple: deadd45e622d3803-simple,
+  simple: f7100025dd3660945f22-simple,
 }
