+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/hash-local-node.txt
@@ -1,0 +1,8 @@
+Object {
+  btn--info_is-disabled_1: _1345eb888e62f079d512-btn--info_is-disabled_1,
+  btn-info_is-disabled: _1345eb888e62f079d512-btn-info_is-disabled,
+  foo: bar,
+  foo_bar: _1345eb888e62f079d512-foo_bar,
+  my-btn-info_is-disabled: value,
+  simple: _1345eb888e62f079d512-simple,
+}
