+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/hash-local.txt
@@ -1,8 +1,8 @@
 Object {
-  btn--info_is-disabled_1: _6fc78548db75122e-btn--info_is-disabled_1,
-  btn-info_is-disabled: _6fc78548db75122e-btn-info_is-disabled,
+  btn--info_is-disabled_1: c3300175e36005009d2d-btn--info_is-disabled_1,
+  btn-info_is-disabled: c3300175e36005009d2d-btn-info_is-disabled,
   foo: bar,
-  foo_bar: _6fc78548db75122e-foo_bar,
+  foo_bar: c3300175e36005009d2d-foo_bar,
   my-btn-info_is-disabled: value,
-  simple: _6fc78548db75122e-simple,
+  simple: c3300175e36005009d2d-simple,
 }
