+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/hash-node.txt
@@ -1,0 +1,8 @@
+Object {
+  btn--info_is-disabled_1: _8dd8d691a7f037527733,
+  btn-info_is-disabled: _8dd8d691a7f037527733,
+  foo: bar,
+  foo_bar: _8dd8d691a7f037527733,
+  my-btn-info_is-disabled: value,
+  simple: _8dd8d691a7f037527733,
+}
