+++ tests/rspack-test/configCases/css/local-ident-name/__snapshot__/unique-name-id-contenthash.txt
@@ -1,8 +1,8 @@
 Object {
-  btn--info_is-disabled_1: -_style_module_css_uniqueName-id-contenthash-[contenthash],
-  btn-info_is-disabled: -_style_module_css_uniqueName-id-contenthash-[contenthash],
+  btn--info_is-disabled_1: -_style_module_css_uniqueName-id-contenthash-66ccbbe4ce365e96ef13,
+  btn-info_is-disabled: -_style_module_css_uniqueName-id-contenthash-66ccbbe4ce365e96ef13,
   foo: bar,
-  foo_bar: -_style_module_css_uniqueName-id-contenthash-[contenthash],
+  foo_bar: -_style_module_css_uniqueName-id-contenthash-66ccbbe4ce365e96ef13,
   my-btn-info_is-disabled: value,
-  simple: -_style_module_css_uniqueName-id-contenthash-[contenthash],
+  simple: -_style_module_css_uniqueName-id-contenthash-66ccbbe4ce365e96ef13,
 }
