+++ tests/rspack-test/defaultsCases/default/base.js
@@ -214,18 +214,27 @@ module.exports = {
 			        esModule: true,
 			        exportsConvention: as-is,
 			        exportsOnly: false,
+			        localIdentHashDigest: base64url,
+			        localIdentHashDigestLength: 6,
+			        localIdentHashFunction: md4,
 			        localIdentName: [fullhash],
 			      },
 			      css/global: Object {
 			        esModule: true,
 			        exportsConvention: as-is,
 			        exportsOnly: false,
+			        localIdentHashDigest: base64url,
+			        localIdentHashDigestLength: 6,
+			        localIdentHashFunction: md4,
 			        localIdentName: [fullhash],
 			      },
 			      css/module: Object {
 			        esModule: true,
 			        exportsConvention: as-is,
 			        exportsOnly: false,
+			        localIdentHashDigest: base64url,
+			        localIdentHashDigestLength: 6,
+			        localIdentHashFunction: md4,
 			        localIdentName: [fullhash],
 			      },
 			      json: Object {
